rename project dirs
