+++ upwork_scraper/profile_scanner.py
@@ -9,9 +9,9 @@ from typing import Optional, Union
 from bs4 import BeautifulSoup, NavigableString, Tag
 from retry import retry
 
-from argyle_upwork.driver import ChromeDriver, DriverManager
-from argyle_upwork.logger import logger
-from argyle_upwork.models.profile import (AccountSection, LocationSection,
+from upwork_scraper.driver import ChromeDriver, DriverManager
+from upwork_scraper.logger import logger
+from upwork_scraper.models.profile import (AccountSection, LocationSection,
                                           Profile, ProfilePage)
